+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.1.15] - 2025-04-03
+
+### Added
+- Changes in scripts/README.md
+- Changes in scripts/checksum.sh
+
 ## [0.1.14] - 2025-04-03
 
 ### Added